+++ demos/world/src/main/java/is/codion/demos/world/ui/WorldAppPanel.java
@@ -41,25 +41,29 @@ import javax.swing.SwingConstants;
 import java.util.List;
 import java.util.Locale;
 
+import static java.util.Collections.emptyList;
+
 public final class WorldAppPanel extends EntityApplicationPanel<WorldAppModel> {
 
-	public WorldAppPanel(WorldAppModel applicationModel) {
-		super(applicationModel);
+	static {
 		FrameworkIcons.instance().add(Foundation.MAP, Foundation.PAGE_EXPORT, Foundation.PAGE_ADD, Foundation.CHECK);
 	}
 
+	public WorldAppPanel(WorldAppModel applicationModel) {
+		super(applicationModel, createPanels(applicationModel), emptyList());
+	}
+
 	// tag::initializeEntityPanels[]
-	@Override
-	protected List<EntityPanel> createEntityPanels() {
+	private static List<EntityPanel> createPanels(WorldAppModel applicationModel) {
 		CountryModel countryModel = (CountryModel)
-						applicationModel().entityModels().get(Country.TYPE);
+						applicationModel.entityModels().get(Country.TYPE);
 		CountryPanel countryPanel = new CountryPanel(countryModel);
 
 		ContinentModel continentModel = (ContinentModel)
-						applicationModel().entityModels().get(Continent.TYPE);
+						applicationModel.entityModels().get(Continent.TYPE);
 		ContinentPanel continentPanel = new ContinentPanel(continentModel);
 
-		SwingEntityModel lookupModel = applicationModel().entityModels().get(Lookup.TYPE);
+		SwingEntityModel lookupModel = applicationModel.entityModels().get(Lookup.TYPE);
 		EntityPanel lookupPanel = new EntityPanel(lookupModel,
 						new LookupTablePanel(lookupModel.tableModel()));